Set TMC_UART::getTime to return 32bit value because size_t is 16bit on AVR

+++ examples/UnitTests/src/uart.cpp
@@ -8,7 +8,7 @@ int TMC_UART::available() {
     return 0;
 }
 
-size_t TMC_UART::getTime() const {
+uint32_t TMC_UART::getTime() const {
     return 0;
 }
